# perf: remove the extra else in propagate