In https://github.com/stackblitz/alien-signals/blob/master/src/system.ts#L138, it will always be executed

```ts
sub.flags = subFlags | targetFlag
```

Therefore, https://github.com/stackblitz/alien-signals/blob/master/src/system.ts#L168 is an additional calculation.

example:
```js
var a = 1 << 1; // 2

console.log((a |= 1 << 1)); // 2
console.log((a |= 1 << 1)); // 2
```

before:
- clk: ~3.93 GHz
- cpu: Apple M3 Max
- runtime: bun 1.1.42 (arm64-darwin)

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | ` 23.19 ns/iter` | ` 22.76 ns` | ` 22.85 ns` | ` 28.74 ns` | ` 66.13 ns` |
| propagate: 1 * 10    | `155.08 ns/iter` | `152.24 ns` | `156.47 ns` | `176.05 ns` | `179.67 ns` |
| propagate: 1 * 100   | `  1.45 µs/iter` | `  1.44 µs` | `  1.45 µs` | `  1.48 µs` | `  1.51 µs` |
| propagate: 10 * 1    | `276.37 ns/iter` | `273.06 ns` | `276.57 ns` | `308.34 ns` | `359.09 ns` |
| propagate: 10 * 10   | `  1.58 µs/iter` | `  1.57 µs` | `  1.59 µs` | `  1.60 µs` | `  1.60 µs` |
| propagate: 10 * 100  | ` 18.83 µs/iter` | ` 18.71 µs` | ` 18.79 µs` | ` 19.14 µs` | ` 19.21 µs` |
| propagate: 100 * 1   | `  2.66 µs/iter` | `  2.66 µs` | `  2.67 µs` | `  2.67 µs` | `  2.68 µs` |
| propagate: 100 * 10  | ` 22.23 µs/iter` | ` 22.02 µs` | ` 22.41 µs` | ` 22.46 µs` | ` 22.54 µs` |
| propagate: 100 * 100 | `178.65 µs/iter` | `174.92 µs` | `178.46 µs` | `200.21 µs` | `209.88 µs` |
| complex: 1 * 1       | ` 49.91 ns/iter` | ` 42.34 ns` | ` 51.82 ns` | ` 92.03 ns` | `334.84 ns` |
| complex: 1 * 10      | `421.49 ns/iter` | `380.45 ns` | `428.93 ns` | `527.75 ns` | `894.21 ns` |
| complex: 1 * 100     | `  5.56 µs/iter` | `  5.49 µs` | `  5.59 µs` | `  5.67 µs` | `  5.74 µs` |
| complex: 10 * 1      | `582.33 ns/iter` | `565.81 ns` | `574.82 ns` | `789.27 ns` | `798.74 ns` |
| complex: 10 * 10     | `  4.78 µs/iter` | `  4.72 µs` | `  4.78 µs` | `  4.85 µs` | `  5.31 µs` |
| complex: 10 * 100    | ` 75.34 µs/iter` | ` 72.17 µs` | ` 75.38 µs` | ` 83.46 µs` | `252.38 µs` |
| complex: 100 * 1     | `  6.65 µs/iter` | `  6.62 µs` | `  6.68 µs` | `  6.69 µs` | `  6.70 µs` |
| complex: 100 * 10    | ` 78.27 µs/iter` | ` 75.13 µs` | ` 78.29 µs` | ` 84.63 µs` | `252.50 µs` |
| complex: 100 * 100   | `778.07 µs/iter` | `712.46 µs` | `796.33 µs` | `  1.17 ms` | `  1.31 ms` |

now:

- clk: ~3.93 GHz
- cpu: Apple M3 Max
- runtime: bun 1.1.42 (arm64-darwin)

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | ` 23.07 ns/iter` | ` 22.67 ns` | ` 22.72 ns` | ` 28.55 ns` | ` 69.70 ns` |
| propagate: 1 * 10    | `153.58 ns/iter` | `151.15 ns` | `154.72 ns` | `172.00 ns` | `181.33 ns` |
| propagate: 1 * 100   | `  1.48 µs/iter` | `  1.47 µs` | `  1.48 µs` | `  1.49 µs` | `  1.51 µs` |
| propagate: 10 * 1    | `276.56 ns/iter` | `272.92 ns` | `276.47 ns` | `307.88 ns` | `363.62 ns` |
| propagate: 10 * 10   | `  1.58 µs/iter` | `  1.57 µs` | `  1.58 µs` | `  1.59 µs` | `  1.59 µs` |
| propagate: 10 * 100  | ` 18.95 µs/iter` | ` 18.81 µs` | ` 19.10 µs` | ` 19.13 µs` | ` 19.17 µs` |
| propagate: 100 * 1   | `  2.66 µs/iter` | `  2.65 µs` | `  2.66 µs` | `  2.68 µs` | `  2.68 µs` |
| propagate: 100 * 10  | ` 21.04 µs/iter` | ` 20.93 µs` | ` 21.05 µs` | ` 21.09 µs` | ` 21.09 µs` |
| propagate: 100 * 100 | `178.94 µs/iter` | `176.42 µs` | `178.75 µs` | `190.75 µs` | `219.29 µs` |
| complex: 1 * 1       | ` 48.26 ns/iter` | ` 42.31 ns` | ` 49.58 ns` | ` 90.36 ns` | `286.89 ns` |
| complex: 1 * 10      | `401.03 ns/iter` | `367.09 ns` | `404.23 ns` | `568.95 ns` | `850.59 ns` |
| complex: 1 * 100     | `  5.59 µs/iter` | `  5.52 µs` | `  5.61 µs` | `  5.66 µs` | `  5.77 µs` |
| complex: 10 * 1      | `711.96 ns/iter` | `693.10 ns` | `706.85 ns` | `922.04 ns` | `945.72 ns` |
| complex: 10 * 10     | `  5.68 µs/iter` | `  5.62 µs` | `  5.70 µs` | `  5.76 µs` | `  5.94 µs` |
| complex: 10 * 100    | ` 74.53 µs/iter` | ` 71.33 µs` | ` 74.54 µs` | ` 86.75 µs` | `257.67 µs` |
| complex: 100 * 1     | `  6.83 µs/iter` | `  6.76 µs` | `  6.86 µs` | `  6.89 µs` | `  6.91 µs` |
| complex: 100 * 10    | ` 76.77 µs/iter` | ` 73.67 µs` | ` 76.75 µs` | ` 88.25 µs` | `251.88 µs` |
| complex: 100 * 100   | `723.24 µs/iter` | `707.71 µs` | `721.79 µs` | `786.92 µs` | `  1.07 ms` |