@johnsoncodehk 2024-12-23
This was intentional in the past to avoid unnecessary flags property assignments as much as possible, you pointed me to the logic that now doesn't avoid all unnecessary assignments.

Since your changes are mostly equivalent (except for the recursive case where the flags property is assigned twice), I'll merge it, thanks!