@medz 2024-12-23
@johnsoncodehk Also, can you share your formatting rules? I use zed which automatically calls prettier to format the code.

Because I don't know your code style rules very well, there will always be unexpected formatting.