@medz 2024-12-23
> This was intentional in the past to avoid unnecessary flags property assignments as much as possible, you pointed me to the logic that now doesn't avoid all unnecessary assignments.
> 
> Since your changes are mostly equivalent (except for the recursive case where the flags property is assigned twice), I'll merge it, thanks!

Yes, they are equivalent. Although it allocates two flags accesses, it avoids logic bifurcation and helps reduce the instruction set. (Although it is almost negligible in dynamic programming languages)

But reducing logic bifurcation and prediction can reduce the CPU usage of most static programming languages ​​(before submitting the PR, I did an experiment in Dart with attribute assignment and if/else execution 10 million times, and the CPU usage was reduced)

So I submitted it.